+++ CHANGELOG.md
@@ -7,4 +7,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Removed
  - Removed old, unused SPlus display interface project
- 
+ - Removing unused VolumeControl property from IDisplayWithAudio